backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt==1.6.1
pydantic
